requirements.txt
flask
google-generativeai
google-auth
google-api-python-client
python-dotenv
